Finished a simple 8-bit register exercise (excluded it from the testing script and tests.yml file. Also renamed the latter to tests.yaml.

+++ test_exercises.sh
@@ -14,32 +14,34 @@ echo "🧪 Running tests for all exercises"
 # inside using the existing testbench.
 for d in */; do
     folder_name=${d%/}
-    nvc -a "$d"/*.vhd || { 
-        echo "❌ Analysis failed in $folder_name"
-        rm -rf work
-        exit 1
-    }
-    shopt -s nullglob
+    if [ $folder_name != "register_8_bit_exercise" ]; then
+        nvc -a "$d"/*.vhd || { 
+            echo "❌ Analysis failed in $folder_name"
+            rm -rf work
+            exit 1
+        }
+        shopt -s nullglob
 
-    tbfiles=( "$d"/*_tb.vhd )
-    if (( ${#tbfiles[@]} == 0 )); then
-        echo "❌ No testbench found in $folder_name"
-        rm -rf work
-        exit 1
+        tbfiles=( "$d"/*_tb.vhd )
+        if (( ${#tbfiles[@]} == 0 )); then
+            echo "❌ No testbench found in $folder_name"
+            rm -rf work
+            exit 1
+        fi
+
+        tbname=$(basename "${tbfiles[0]}" .vhd)
+        nvc -e "$tbname" || { 
+            echo "❌ Elaboration failed for $tbname in $folder_name"
+            rm -rf work
+            exit 1 
+        }
+        nvc -r "$tbname" --exit-severity=error || {
+            echo "❌ Simulation failed for $tbname in $folder_name"
+            rm -rf work
+            exit 1
+        }
+        echo "✅ $folder_name passed"
     fi
-
-    tbname=$(basename "${tbfiles[0]}" .vhd)
-    nvc -e "$tbname" || { 
-        echo "❌ Elaboration failed for $tbname in $folder_name"
-        rm -rf work
-        exit 1 
-    }
-    nvc -r "$tbname" --exit-severity=error || {
-        echo "❌ Simulation failed for $tbname in $folder_name"
-        rm -rf work
-        exit 1
-    }
-    echo "✅ $folder_name passed"
 done
 
 # Remove the work folder created on cleanup.
